+++ src/pages/index.js
@@ -1,6 +1,7 @@
 import GamePagesEnum from '../constants/gamePagesEnum';
 import BottleRocketPage from './events/bottleRocket';
 import HomePage from './home';
+import InventoryPage from './inventory';
 import ItemPage from './item';
 import KitchenPage from './kitchen';
 import NPCSPage from './npcs';
@@ -10,9 +11,11 @@ import QuestsPage from './quests';
 import FarmSupplyPage from './town/supply';
 import VaultPage from './town/vault';
 import WellPage from './town/well';
+import MuseumPage from './town/museum';
 
 const Pages = {
     [GamePagesEnum.HOME]: new HomePage,
+    [GamePagesEnum.INVENTORY]: new InventoryPage,
     [GamePagesEnum.KITCHEN]: new KitchenPage,
     [GamePagesEnum.QUESTS]: new QuestsPage,
     [GamePagesEnum.NPCS]: new NPCSPage,
@@ -24,6 +27,7 @@ const Pages = {
     [GamePagesEnum.FARM_SUPPLY]: new FarmSupplyPage,
     [GamePagesEnum.VAULT]: new VaultPage,
     [GamePagesEnum.WELL]: new WellPage,
+    [GamePagesEnum.MUSEUM]: new MuseumPage,
 
     [GamePagesEnum.BOTTLE_ROCKET]: new BottleRocketPage,
 };